moved instances to Internal.hs. Also upped version to 1.0.0, since we export Parser now from Aeson.Safe
